\name{ReplicatedPointPatterns-package}
\alias{ReplicatedPointPatterns-package}
\alias{ReplicatedPointPatterns}
\docType{package}
\title{
  ReplicatedPointPatterns
}
\description{
  This package analyses replicated point pattern data and links variation among the second-order structure of point patterns to covariates. Functions are available for fitting both linear models and mixed-effects models. 
}
\details{
  \tabular{ll}{
    Package: \tab ReplicatedPointPatterns\cr
    Type: \tab Package\cr
    Version: \tab 0.2 \cr
    Date: \tab 2014-10-17\cr
    License: \tab GPL-2\cr
  }
  ~~ An overview of how to use the package, including the most important ~~
  ~~ functions ~~
}
\author{
  Robert Bagchi
  Maintainer: Robert Bagchi <robert.bagchi@uconn.edu>
}
\references{
  ~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
  ~~ Optional links to other man pages, e.g. ~~
  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
